+++ modules/onboarding/versions.tf
@@ -6,5 +6,9 @@ terraform {
       source  = "hashicorp/azuread"
       version = "~> 3.0"
     }
+    azapi = {
+      source  = "azure/azapi"
+      version = "~> 2.0"
+    }
   }
 }